cast to fix logging.

+++ Pulp/Network/Reliability/ReliabilityLayer.cpp
@@ -1271,7 +1271,7 @@ void ReliabilityLayer::update(core::FixedBitStreamBase& bs, NetSocket& socket, S
 
 		X_ASSERT(bs.size() == dataGramHdrSizeBits(), "Invalid size logic")(bs.size(), dataGramHdrSizeBits());
 
-		X_LOG0_IF(vars_.debugEnabled(), "NetRel", "Sending DataGram number: ^5%" PRIu16 "^7 numPackets: ^5%" PRIuS, dgh.number, end - begin);
+		X_LOG0_IF(vars_.debugEnabled(), "NetRel", "Sending DataGram number: ^5%" PRIu16 "^7 numPackets: ^5%" PRIuS, dgh.number, static_cast<size_t>(end - begin));
 
 		DataGramHistory* pHistory = createDataGramHistory(dgh.number, time);
